Replaced links to MDN
I replaced the link in the "Getting Started" section and changed the wording. I changed the heading of the "Extension/Add-On Development" section to just say "Add-Ons." I changed the link in that section to the thunderbird.net add-ons page and changed the wording. Under Documentation I changed the two spots that referenced MDN. I replaced their links for the correct ones and changed the wording.