Add latest version to data attribute on download pages.
I'm working on adding support for Thunderbird to Pollbot (https://github.com/mozilla/PollBot)
with the goal of having something similar to https://mozilla.github.io/delivery-dashboard/.

One feature of Pollbot is it scrapes the latest Firefox version out of the various download
pages on www.mozilla.org in order to check that it matches what's provided by product-details.
The easiest way to do that is to have a data attribute somewhere in the page. To make it
easier to write the query selector, I added the id attribute on the element as well.